requirements.txt
# requirements.txt
Django>=4.2,<5.0
Pillow>=9.0,<10.0
django-crispy-forms>=2.0,<3.0
crispy-bootstrap5>=0.7,<1.0
gunicorn>=20.0,<21.0
whitenoise>=6.0,<7.0